+++ docs/guides/communication-between-modules.md
@@ -15,7 +15,7 @@ GraphQL Modules provide multiple ways of doing it, and you should pick the right
 If your modules are coupled and you they should know each other because their logic and combined with each user, you cam use direct dependency injection.
 
 ```typescript
-import { Injectable } from '@graphql-modules/core';
+import { Injectable } from '@graphql-modules/di';
 import { OtherProvider } from '../my-other-module/other.provider';
 
 @Injectable()
@@ -31,7 +31,7 @@ export class MyProvider {
 If you wish to communicate other module with importing directly from it, you can use dependency-injection tokens:
 
 ```typescript
-import { Inject, Injectable } from '@graphql-modules/core';
+import { Inject, Injectable } from '@graphql-modules/di';
 
 export interface IOtherProviderSignature {
     doSomething: () => void;
@@ -118,7 +118,8 @@ export const BarModule = new GraphQLModule({
 Then, to use `CommunicationBridge`, you can do the following:
 
 ```typescript
-import { CommunicationBridge, Injectable } from '@graphql-modules/core';
+import { Injectable } from '@graphql-modules/di';
+import { CommunicationBridge } from '@graphql-modules/core';
 
 @Injectable()
 export class MyProvider {
