+++ examples/basic-with-dependency-injection/src/modules/blog/index.ts
@@ -1,5 +1,5 @@
-import {GraphQLModule} from '@graphql-modules/core';
-import {Blog} from './providers/blog';
+import { GraphQLModule } from '@graphql-modules/core';
+import { Blog } from './providers/blog';
 import gql from 'graphql-tag';
 import resolvers from './resolvers';
 import { UserModule } from '../user';
