+++ examples/basic/src/index.ts
@@ -1,4 +1,3 @@
-import 'reflect-metadata';
 import { AppModule } from './app/app.module';
 import { bootstrap } from './server';
