+++ generate-api-docs.sh
@@ -4,6 +4,7 @@ typedoc --module commonjs --excludeNotExported --excludePrivate --ignoreCompiler
 typedoc --module commonjs --excludeNotExported --excludePrivate --ignoreCompilerErrors --mode file --theme markdown --out 'docs/api/epoxy' 'packages/epoxy/src'
 typedoc --module commonjs --excludeNotExported --excludePrivate --ignoreCompilerErrors --mode file --theme markdown --out 'docs/api/sonar' 'packages/sonar/src'
 typedoc --module commonjs --excludeNotExported --excludePrivate --ignoreCompilerErrors --mode file --theme markdown --out 'docs/api/logger' 'packages/logger/src'
+typedoc --module commonjs --excludeNotExported --excludePrivate --ignoreCompilerErrors --mode file --theme markdown --out 'docs/api/di' 'packages/di/src'
 
 yarn replace-in-file "/\.md\)/g" ")" "./docs/api/**/*.md" --isRegex
 yarn replace-in-file "/\.md#/g" "#" "./docs/api/**/*.md" --isRegex
