+++ packages/core/package.json
@@ -27,13 +27,9 @@
       "node"
     ]
   },
-  "peerDependencies": {
-    "reflect-metadata": "0.1.12"
-  },
   "dependencies": {
     "@graphql-modules/epoxy": "0.2.0",
-    "@graphql-modules/logger": "0.2.0",
-    "@graphql-modules/sonar": "0.2.0",
+    "@graphql-modules/di": "0.2.0",
     "dependency-graph": "0.7.2",
     "graphql-tools": "4.0.3",
     "tslib": "1.9.3"
@@ -41,7 +37,6 @@
   "devDependencies": {
     "graphql-tag": "2.10.0",
     "jest": "23.6.0",
-    "reflect-metadata": "0.1.12",
     "tslint": "5.11.0",
     "typescript": "3.1.6"
   },
