--- packages/core/src/di/index.ts
@@ -1,6 +1,0 @@
-export { __decorate as decorate } from 'tslib';
-export { Inject } from './inject';
-export { Injectable } from './injectable';
-export { Injector } from './injector';
-export * from './types';
-export * from './utils';
