+++ packages/core/src/errors/index.ts
@@ -1,7 +1,4 @@
-export { DependencyProviderNotFoundError } from './dependency-provider-not-found';
-export { ServiceIdentifierNotFoundError } from './service-identifier-not-found';
 export { DependencyModuleNotFoundError } from './dependency-module-not-found';
 export { SchemaNotValidError } from './schema-not-valid';
 export { DependencyModuleUndefinedError } from './dependency-module-undefined';
 export { TypeDefNotFoundError } from './typedef-not-found';
-export { ProviderNotValidError } from './provider-not-valid';
