+++ packages/core/src/index.ts
@@ -1,5 +1,5 @@
 export * from './graphql-module';
 export * from './communication';
 export * from './resolvers-composition';
-export * from './di';
 export * from './errors';
+export * from './types';
