+++ packages/di/.gitignore
@@ -1,0 +1,6 @@
+node_modules
+.idea
+dist
+build
+out
+temp
