+++ packages/di/.npmignore
@@ -1,0 +1,5 @@
+src
+node_modules
+tests
+!dist
+tsconfig.json
