+++ packages/di/src/errors/index.ts
@@ -1,0 +1,4 @@
+export * from './dependency-provider-not-found';
+export * from './provider-not-valid';
+export * from './service-identifier-not-found';
+export * from './provider-already-defined';
