+++ packages/di/src/index.ts
@@ -1,0 +1,8 @@
+// tslint:disable-next-line:no-reference
+/// <reference path="../../../node_modules/reflect-metadata/index.d.ts" />
+export * from './inject';
+export * from './injectable';
+export * from './injector';
+export * from './types';
+export * from './utils';
+export * from './errors';
