+++ packages/di/src/injectable.ts
@@ -1,7 +1,5 @@
 import { DESIGN_PARAM_TYPES } from './utils';
 
-declare var Reflect: any;
-
 export function Injectable() {
   return (target: any) => {
     if (!Reflect.hasMetadata(DESIGN_PARAM_TYPES, target)) {
