+++ packages/di/tsconfig.json
