+++ packages/epoxy/tsconfig.json
@@ -2,9 +2,8 @@
   "compilerOptions": {
     "experimentalDecorators": true,
     "module": "commonjs",
-    "target": "es5",
-    "lib": ["es6", "esnext", "es2015"],
-    "noImplicitAny": false,
+    "target": "es2015",
+    "lib": ["esnext"],
     "suppressImplicitAnyIndexErrors": true,
     "moduleResolution": "node",
     "emitDecoratorMetadata": true,
