+++ packages/logger/tsconfig.json
@@ -2,8 +2,8 @@
   "compilerOptions": {
     "experimentalDecorators": true,
     "module": "commonjs",
-    "target": "es5",
-    "lib": ["es6", "esnext", "es2015"],
+    "target": "es2015",
+    "lib": ["esnext"],
     "noImplicitAny": false,
     "suppressImplicitAnyIndexErrors": true,
     "moduleResolution": "node",
