+++ website/sidebars.json
@@ -4,7 +4,8 @@
       "api/core/api-readme",
       "api/sonar/api-readme",
       "api/epoxy/api-readme",
-      "api/logger/api-readme"
+      "api/logger/api-readme",
+      "api/di/api-readme"
     ]
   },
   "docs": {
